test_setup.py
import os
from dotenv import load_dotenv
from langchain_groq import ChatGroq

load_dotenv()

llm = ChatGroq(model="llama-3.3-70b-versatile", temperature=0)
response = llm.invoke("Say 'Hello from VS Code agent project!' in pirate speak.")
print(response.content)